more auth cleanup
