+++ routes/genres.js
@@ -2,7 +2,7 @@ const { Genre, validate } = require("../models/genre");
 const mongoose = require("mongoose");
 const express = require("express");
 const router = express.Router();
-const auth = require("../middleware/auth");
+const auth = require("../controllers/auth");
 const admin = require("../middleware/admin");
 
 // const genres = [
