+++ routes/mail_routes.js
@@ -3,12 +3,6 @@ const router = express.Router();
 const mailController = require("../controllers/mail");
 const mailMiddleware = require("../middleware/mail");
 
-router.post("/send", mailController.sendgrid, mailMiddleware.sendgrid);
-router.post(
-  "/register-confirm",
-  mailController.sendgrid,
-  mailMiddleware.sendgrid
-);
 router.post(
   "/send-contact",
   mailController.sendContact,
