+++ web/src/api/authApi.ts
@@ -11,7 +11,7 @@ export const logout = async (
   request: LogoutRequest
 ): Promise<ApiResponse<User>> => {
   // console.log("authApi.logout()");
-  const client = new ApiClient<User>("/auth/logout");
+  const client = new ApiClient<User>("/user/logout");
   try {
     return client.post(request._id).then((response) => {
       return Promise.resolve(response);
@@ -25,7 +25,7 @@ export const register = async (
   request: RegisterRequest
 ): Promise<ApiResponse<User>> => {
   // console.log("authService.register");
-  const client = new ApiClient<User>("auth/sign-up");
+  const client = new ApiClient<User>("user/sign-up");
   try {
     return client.post(request).then((response) => {
       return Promise.resolve(response);
@@ -38,7 +38,7 @@ export const register = async (
 export const login = async (
   request: LoginRequest
 ): Promise<ApiResponse<User>> => {
-  const client = new ApiClient<User>("auth/login");
+  const client = new ApiClient<User>("user/login");
   try {
     return client.post(request).then((response) => {
       // console.log("authApi.login response: ", response);
